medicart/requirements.txt
asgiref==3.9.2
Django==4.2.25
django-cors-headers==4.9.0
django-rest-passwordreset==1.5.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
